[JAVA-33592] Created new module core-java-string-operations-9 and moved code from core-java-string-operations-7
