+++ core-java-modules/core-java-string-operations-6/README.md
@@ -11,3 +11,4 @@
 - [Check if a String Has All Unique Characters in Java](https://www.baeldung.com/java-check-string-all-unique-chars)
 - [Performance Comparison Between Different Java String Concatenation Methods](https://www.baeldung.com/java-string-concatenation-methods)
 - [Replacing Single Quote with \’ in Java String](https://www.baeldung.com/java-replacing-single-quote-string)
+- More articles: [[<-- prev]](../core-java-string-operations-5) [[next -->]](../core-java-string-operations-7)
