+++ core-java-modules/core-java-string-operations-7/README.md
@@ -2,17 +2,12 @@
 ### Relevant Articles:
 - [How to Center Text Output in Java](https://www.baeldung.com/java-center-text-output)
 - [Capitalize the First Letter of Each Word in a String](https://www.baeldung.com/java-string-initial-capital-letter-every-word)
-- [Check if a String Contains Only Unicode Letters](https://www.baeldung.com/java-string-all-unicode-characters)
 - [Create a “Mutable” String in Java](https://www.baeldung.com/java-mutable-string)
 - [Check if a String Contains a Number Value in Java](https://www.baeldung.com/java-string-number-presence)
-- [Difference Between String isEmpty() and isBlank()](https://www.baeldung.com/java-string-isempty-vs-isblank)
 - [String’s Maximum Length in Java](https://www.baeldung.com/java-strings-maximum-length)
 - [Java’s String.length() and String.getBytes().length](https://www.baeldung.com/java-string-length-vs-getbytes-length)
-- [Replace Non-Printable Unicode Characters in Java](https://www.baeldung.com/java-replace-non-printable-unicode-characters)
 - [Check If a Java StringBuilder Object Contains a Character](https://www.baeldung.com/java-check-stringbuilder-object-contains-character)
 - [Comparing One String With Multiple Values in One Expression in Java](https://www.baeldung.com/java-compare-string-multiple-values-one-expression)
-- [UTF-8 Validation in Java](https://www.baeldung.com/java-utf-8-validation)
-- [Simple Morse Code Translation in Java](https://www.baeldung.com/java-morse-code-english-translate)
-- [How to Determine if a String Contains Invalid Encoded Characters](https://www.baeldung.com/java-check-string-contains-invalid-encoded-characters)
 - [Regular Expression for Password Validation in Java](https://www.baeldung.com/java-regex-password-validation)
 - [Mask an Email Address and Phone Number in Java](https://www.baeldung.com/java-mask-email-address-phone-number)
+- More articles: [[<-- prev]](../core-java-string-operations-6) [[next -->]](../core-java-string-operations-8)
