+++ core-java-modules/core-java-string-operations-9/README.md
@@ -1,0 +1,9 @@
+
+### Relevant Articles:
+- [Check if a String Contains Only Unicode Letters](https://www.baeldung.com/java-string-all-unicode-characters)
+- [Difference Between String isEmpty() and isBlank()](https://www.baeldung.com/java-string-isempty-vs-isblank)
+- [Replace Non-Printable Unicode Characters in Java](https://www.baeldung.com/java-replace-non-printable-unicode-characters)
+- [UTF-8 Validation in Java](https://www.baeldung.com/java-utf-8-validation)
+- [Simple Morse Code Translation in Java](https://www.baeldung.com/java-morse-code-english-translate)
+- [How to Determine if a String Contains Invalid Encoded Characters](https://www.baeldung.com/java-check-string-contains-invalid-encoded-characters)
+- More articles: [[<-- prev]](../core-java-string-operations-8)
