+++ core-java-modules/core-java-string-operations-9/pom.xml
@@ -3,9 +3,9 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>core-java-string-operations-7</artifactId>
+    <artifactId>core-java-string-operations-9</artifactId>
     <packaging>jar</packaging>
-    <name>core-java-string-operations-7</name>
+    <name>core-java-string-operations-9</name>
 
     <parent>
         <groupId>com.baeldung.core-java-modules</groupId>
@@ -23,6 +23,21 @@
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-text</artifactId>
             <version>${commons-text.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.tika</groupId>
+            <artifactId>tika-core</artifactId>
+            <version>${apache.tika.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.tika</groupId>
+            <artifactId>tika-parsers-standard-package</artifactId>
+            <version>${apache.tika.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.ibm.icu</groupId>
+            <artifactId>icu4j</artifactId>
+            <version>${icu4j.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.commons</groupId>
@@ -45,7 +60,6 @@
     </build>
 
     <properties>
-
         <apache.commons.lang3.version>3.13.0</apache.commons.lang3.version>
         <apache.commons.collection.version>4.4</apache.commons.collection.version>        
         <apache.tika.version>2.9.1</apache.tika.version>
