+++ core-java-modules/core-java-string-operations-9/src/main/java/com/baeldung/unicode/UnicodeLetterChecker.java
