+++ core-java-modules/core-java-string-operations-9/src/test/java/com/baeldung/unicode/UnicodeLetterCheckerUnitTest.java
