+++ core-java-modules/core-java-string-operations-9/src/test/java/com/baeldung/utf8validation/UTF8ValidationUnitTest.java
