+++ core-java-modules/pom.xml
@@ -204,6 +204,7 @@
         <module>core-java-string-operations-6</module>
         <module>core-java-string-operations-7</module>
         <module>core-java-string-operations-8</module>
+        <module>core-java-string-operations-9</module>
         <module>core-java-regex</module>
         <module>core-java-regex-2</module>
         <module>core-java-regex-3</module>
